more code for opserver, ophashtable and queue

+++ opserver.h
@@ -9,7 +9,7 @@ struct opserver {
   uint32_t       id;
   pthread_t      tid;
   size_t         maxmsgsize;
-  size_t         qlen;
+  int            qlen;
   opserver_func *func;
   UT_hash_handle hh;
   struct queue   queue;
@@ -29,7 +29,7 @@ static inline pthread_t opserver_getTid(opserver_t this) {
 }
 
 // new operator 
-extern opserver_t opserver_new(int id, opserver_func *func,
+extern opserver_t opserver_new(uint32_t id, opserver_func *func,
 			       size_t maxmsgsize, int qlen);
 
 // core mothods
